### 2023-05-22 #5488 @editorialbot
[PRE REVIEW]: ClassipyGRB: Machine Learning-Based Classification and Visualization of Gamma Ray Bursts using t-SNE
**Submitting author:** @KenethGarcia (<a href="http://orcid.org/0009-0001-2607-6359">Keneth Stiven Garcia Cifuentes</a>)
**Repository:** https://github.com/KenethGarcia/ClassiPyGRB
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
