### 2023-05-22 #5488 @dfm
@KenethGarcia — Thanks for your submission! I have submitted this submission for "scope review" by the JOSS editorial board to consider whether or not the code is within publication scope for JOSS. This process typically takes a week or two and I'll keep you updated on the progress.

In the meantime, please fix the syntax of your manuscript which is currently inconsistent with the docs [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). You can check that the paper compiles following the instructions [here](https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles).
