+++ JOSS_Docs/paper.bib
@@ -1,7 +1,7 @@
-@ARTICLE{Steinhardt2023,
+@ARTICLE{Steinhardt:2023,
        author = {{Steinhardt}, Charles L. and {Mann}, William J. and {Rusakov}, Vadim and {Jespersen}, Christian K.},
         title = "{Classification of BATSE, Swift, and Fermi Gamma-Ray Bursts from Prompt Emission Alone}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Gamma-ray bursters, Gamma-ray bursts, Dimensionality reduction, Classification systems, 1878, 629, 1943, 253, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2023,
         month = mar,
@@ -17,62 +17,70 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{Pearson:2017,
-  	url = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-  	Archiveprefix = {arXiv},
-  	Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-  	Eprint = {1703.04627},
-  	Journal = {ArXiv e-prints},
-  	Keywords = {Astrophysics - Astrophysics of Galaxies},
-  	Month = mar,
-  	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-  	Year = 2017
+@ARTICLE{Kouveliotou:1993,
+       author = {{Kouveliotou}, Chryssa and {Meegan}, Charles A. and {Fishman}, Gerald J. and {Bhat}, Narayana P. and {Briggs}, Michael S. and {Koshut}, Thomas M. and {Paciesas}, William S. and {Pendleton}, Geoffrey N.},
+        title = "{Identification of Two Classes of Gamma-Ray Bursts}",
+      journal = {Astrophysical Journal Letters},
+     keywords = {Astronomical Catalogs, Gamma Ray Bursts, Gamma Ray Observatory, Frequency Distribution, Spatial Distribution, Transient Response, Space Radiation, GAMMA RAYS: BURSTS},
+         year = 1993,
+        month = aug,
+       volume = {413},
+        pages = {L101},
+          doi = {10.1086/186969},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1993ApJ...413L.101K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@book{Binney:2008,
-  	url = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
-  	Author = {{Binney}, J. and {Tremaine}, S.},
-  	Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
-  	Publisher = {Princeton University Press},
-  	Title = {{Galactic Dynamics: Second Edition}},
-  	Year = 2008
+@ARTICLE{Yang:2022,
+       author = {{Yang}, Jun and {Ai}, Shunke and {Zhang}, Bin-Bin and {Zhang}, Bing and {Liu}, Zi-Ke and {Wang}, Xiangyu Ivy and {Yang}, Yu-Han and {Yin}, Yi-Han and {Li}, Ye and {L{\"u}}, Hou-Jun},
+        title = "{A long-duration gamma-ray burst with a peculiar origin}",
+      journal = {Nature},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2022,
+        month = dec,
+       volume = {612},
+       number = {7939},
+        pages = {232-235},
[…]
