+++ JOSS_Docs/paper.bib
@@ -84,3 +84,20 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023arXiv230408666G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Sanchez-Alarcon:2022,
+       author = {{Sanchez-Alarcon}, Pablo M and {Ascasibar Sequeiros}, Yago},
+        title = "{Fully Adaptive Bayesian Algorithm for Data Analysis, FABADA}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics, Computer Science - Computer Vision and Pattern Recognition, Physics - Data Analysis, Statistics and Probability},
+         year = 2022,
+        month = jan,
+          eid = {arXiv:2201.05145},
+        pages = {arXiv:2201.05145},
+          doi = {10.48550/arXiv.2201.05145},
+archivePrefix = {arXiv},
+       eprint = {2201.05145},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220105145S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
