### 2023-05-22 #5488 @KenethGarcia
First of all, thank you for your answer and comments @dfm

The paper.md was fixed and now works properly using the guidelines from the [Open Journals PDF Generator](https://github.com/marketplace/actions/open-journals-pdf-generator). Now there is a new folder `.github/workflows` in the repository. Also, there is a new updated pre-release [v1.0.0-beta](https://github.com/KenethGarcia/ClassiPyGRB/tree/v1.0.0-beta) with the changes made.

Let me know if there are any other comment or change needed.
