### 2023-07-04 #5488 @KenethGarcia
Dear @dfm

I hope this message finds you well. I am writing to kindly request an update on the progress of the first review. We understand that this process typically takes a week or two; however, it has now been six weeks without any further communication. We would greatly appreciate any information you can provide regarding the status of our submission or if there something that we could do in the meantime. Thank you for your attention, and we look forward to hearing from you soon.

Best regards
