### 2023-07-07 #5488 @arfon
Sure, the editorial board is listed here: https://joss.theoj.org/about#editorial_board
