+++ pyproject.toml
@@ -47,8 +47,7 @@ dependencies = [
     "numpy==1.24.2",
     "openTSNE==0.7.1",
     "pandas==2.0.0",
-    "requests==2.29.0",
-    "scikit_learn==1.2.2",
+    "requests==2.31.0",
     "scipy==1.10.1",
     "tqdm==4.65.0",
     "tables==3.8.0",
