### 2023-10-03 #5923 @editorialbot
[REVIEW]: ClassipyGRB: Machine Learning-Based Classification and Visualization of Gamma Ray Bursts using t-SNE
**Submitting author:** @KenethGarcia (<a href="http://orcid.org/0009-0001-2607-6359">Keneth Stiven Garcia Cifuentes</a>)
**Repository:** https://github.com/KenethGarcia/ClassiPyGRB
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
