### 2023-10-29 #5923 @wkerzendorf
**Functionality**

Installing the software worked but subsequently running the examples did not:

```python
FileNotFoundError                         Traceback (most recent call last)
Cell In[1], line 3
      1 from ClassiPyGRB import SWIFT
      2 swift = SWIFT(res=64)
----> 3 df = swift.obtain_data(name='GRB211211A')

File ~/miniconda/envs/joss_classipygrb/lib/python3.8/site-packages/ClassiPyGRB/_swift.py:174, in SWIFT.obtain_data(self, name, check_disk)
    172         raise RuntimeError(f"Error from tables when trying to read: {e}. Try to re-install tables package.")
    173 else:
--> 174     with resources.open_text(summary_tables, 'summary_general.txt') as file:
    175         grb_names, ids = np.genfromtxt(file, delimiter="|", dtype=str, usecols=(0, 1), unpack=True,
    176                                        autostrip=True)
    177     if len(grb_names) == 0 or not isinstance(grb_names, (Sequence, Mapping, np.ndarray)):

File ~/miniconda/envs/joss_classipygrb/lib/python3.8/importlib/resources.py:126, in open_text(package, resource, encoding, errors)
    124 if reader is not None:
    125     return TextIOWrapper(reader.open_resource(resource), encoding, errors)
--> 126 _check_location(package)
    127 absolute_package_path = os.path.abspath(package.__spec__.origin)
    128 package_path = os.path.dirname(absolute_package_path)

File ~/miniconda/envs/joss_classipygrb/lib/python3.8/importlib/resources.py:82, in _check_location(package)
     80 def _check_location(package):
     81     if package.__spec__.origin is None or not package.__spec__.has_location:
---> 82         raise FileNotFoundError(f'Package has no location {package!r}')

FileNotFoundError: Package has no location <module 'ClassiPyGRB.summary_tables' (namespace)>
```
There are also no automated ways to run the tests or descriptions of how to run tests. There are also no CI/CD workflows for this repository.

The repository has a rudimentary contribution guidelines but without testing frameworks it might be hard to check if contributions break existing code.

**Documentation**
The installation instructions are sparse and there is conflicting information in the paper compared to the main branch. The installation works as described using `pip install .` but it would be nice to give a bit more guidance for users on other systems (macos/windows/other distros. Perhaps giving a conda environment might be helpful.

The Documentation also only exists of a single `Readme.md` without a proper documentation page that would give more insights on API and usage.

**Paper**

The paper has several typos and I would suggest re-reading it carefully and employing a grammar checker.

The paper does not make a complete **State of the Field** analysis but claims that _have not previously been identified by other groups_ without giving specific references.

The paper also often uses qualitative words such as _easy _ and _fast_ without quantitative description of what makes it easy and or fast (these are just some examples).
