### 2023-11-07 KenethGarcia/ClassiPyGRB#1 @dfm
JOSS: Installation instructions are platform specific and incomplete
This code isn't Linux specific so it seems somewhat limiting to only include `apt-get` installation instructions for dependencies.

One more cross platform option would be conda/mamba. The steps that worked for me:

```bash
mamba create -n ClassiPyGRB -c conda-forge python=3.10 pytables=3.8.0 pillow
mamba activate ClassiPyGRB
pip install .
```

Note that the correct version of PyTables needs to be installed with conda/mamba if you don't have the appropriate hdf5 dev library, which is not currently listed as a dependency.

I'd also argue that it's probably unnecessary to pin all the dependency versions (rather than just bounding them). But if there's a good argument for pinning them like that, that should be clearly stated in the README.

ref: https://github.com/openjournals/joss-reviews/issues/5923
