### 2023-11-07 KenethGarcia/ClassiPyGRB#2 @dfm
JOSS: `parallel_lc_limiter` is extremely slow
In [the 3rd tutorial](https://github.com/KenethGarcia/ClassiPyGRB/blob/64305525c9fae77be8d9780ddbf0a3e06aeca9c0/docs/3.BAT_Preprocess.ipynb), it looks like the following line is expected to take approximately 9 seconds:

```python
limited_data, GRB_names, errors = swift.parallel_lc_limiter(GRB_names)
print(errors)
```

But on my computer is takes well over an hour (with some fraction of that time seeming to be spent downloading data). This should be more clearly explained in the tutorial.

ref: https://github.com/openjournals/joss-reviews/issues/5923
