### 2023-11-07 KenethGarcia/ClassiPyGRB#3 @dfm
JOSS: The documentation does not provide a sufficient statement of need
The only documentation is currently the README and the tutorial notebooks. To satisfy the JOSS requirements, the README should explain clearly why this code exists and who the target audience is (a "statement of need").

ref: https://github.com/openjournals/joss-reviews/issues/5923
