### 2023-11-07 KenethGarcia/ClassiPyGRB#4 @dfm
JOSS: The API is not sufficiently well documented
While the tutorials are excellent and thorough, there should also be a clearer discussion of the API (i.e. what functions are available, their arguments, etc.)

ref: https://github.com/openjournals/joss-reviews/issues/5923
