### 2023-11-07 KenethGarcia/ClassiPyGRB#5 @dfm
JOSS: It's not clear how to run the tests
There appear to be tests implemented in the `tests` directory, but it's not clear how to run them and they're not executed on CI. I'd recommend documenting the test running procedure and adding a GitHub Actions workflow to run the tests.

ref: https://github.com/openjournals/joss-reviews/issues/5923
