+++ src/ClassiPyGRB/summary_tables/summary_general.txt
@@ -21,6 +21,57 @@
 #####################################################################################
 #####################################################################################
 ## GRBname | Trig_ID | Trig_time_met |       Trig_time_UTC        |  RA_ground   |  DEC_ground   | Image_position_err |   Image_SNR  |     T90      |   T90_err    |     T50      |   T50_err    | Evt_start_sincetrig | Evt_stop_sincetrig | pcode  | Trigger_method | XRT_detection | comment
+GRB230911C   |   1191188    | 716114877.504 | 2023-09-11T08:47:25.515160 | 2.268013e+02 | -4.135260e+01 |    1.869955e+00    | 1.243908e+01 | 1.448800e+01 | 2.858812e+00 | 6.356000e+00 | 3.067904e+00 |         -240        |         963       | 0.9531 | rate trigger | Yes |    
+GRB230903A   |   1189514    | 715454608.192 | 2023-09-03T17:22:56.248260 | 9.897540e+00 | -4.090888e+01 |    1.853172e+00    | 1.260032e+01 | 2.536000e+00 | 2.694290e-01 | 1.580000e+00 | 5.442499e-01 |         -344        |         963       | 0.4062 | rate trigger | Yes |    
+GRB230826A   |   1187463    | 714771198.976 | 2023-08-26T19:32:47.078920 | 8.304964e+01 |  6.612486e+01 |    6.182241e-01    | 6.046184e+01 | 4.084000e+01 | 1.658322e+00 | 1.218800e+01 | 8.655725e-01 |         -240        |         1091       | 0.8750 | rate trigger | Yes |    
+GRB230818A   |   1186032    | 714094086.208 | 2023-08-18T23:27:34.357180 | 2.858869e+02 |  4.088654e+01 |    1.157656e+00    | 2.467691e+01 | 6.400000e+01 | 3.577709e+01 | 1.600000e+01 | 1.600000e+01 |         -239        |         1041       | 0.3242 | rate trigger | Yes |    
+GRB230816A   |   1185673    | 713894960.128 | 2023-08-16T16:08:48.290780 | 3.284532e+02 |  3.785778e+01 |    1.206236e+00    | 2.326947e+01 | 1.944800e+01 | 3.218161e+00 | 7.256000e+00 | 1.547462e+00 |         -239        |         963       | 0.9844 | rate trigger | Yes |    
+GRB230815A   |   1185505    | 713789427.776 | 2023-08-15T10:49:55.945980 | 1.847605e+02 | -5.806694e+01 |    1.211139e+00    | 2.313502e+01 | 1.700000e+01 | 7.615773e+00 | 2.000000e+00 | 1.000000e+00 |         -240        |         962       | 0.1113 | rate trigger | Yes |    
+GRB230805B   |   1183217    | 712927462.848 | 2023-08-05T11:23:51.076980 | 2.077497e+02 |  3.119459e+01 |    1.904342e+00    | 1.211945e+01 | 1.145200e+02 | 2.334909e+01 | 3.853600e+01 | 1.859213e+01 |         -240        |         962       | 0.4180 | rate trigger | Yes |    
+GRB230802A   |   1182085    | 712651857.920 | 2023-08-02T06:50:26.167840 | 7.622385e+01 |  1.955145e+01 |    7.472170e-01    | 4.612206e+01 | 1.751520e+02 | 1.729387e+01 | 9.946000e+01 | 5.211548e+01 |         -240        |         962       | 0.9922 | rate trigger | Yes |    
[…]
