+++ src/ClassiPyGRB/summary_tables/summary_burst_durations.txt
@@ -14,6 +14,57 @@
 #####################################################################################
 #####################################################################################
 ## GRBname | Trig_ID | Trig_time_met | T100_start | T100_stop | T90_start | T90_stop | T50_start | T50_stop | 1s_peak_start | 1s_peak_stop 
+GRB230911C   |   1191188    | 716114877.504000 | -8.632000 | 9.584000 | -7.376000 | 7.112000 | -2.656000 | 3.700000 | 2.072000 | 3.072000 |
+GRB230903A   |   1189514    | 715454608.192000 | 0.196000 | 2.932000 | 0.276000 | 2.812000 | 0.680000 | 2.260000 | 0.496000 | 1.496000 |
+GRB230826A   |   1187463    | 714771198.976000 | -6.312000 | 47.480000 | -2.752000 | 38.088000 | 1.956000 | 14.144000 | 2.436000 | 3.436000 |
+GRB230818A   |   1186032    | 714094086.208000 | 0.792000 | 80.792000 | 0.792000 | 64.792000 | 0.792000 | 16.792000 | 8.292000 | 9.292000 |
+GRB230816A   |   1185673    | 713894960.128000 | -5.564000 | 16.484000 | -4.248000 | 15.200000 | -0.724000 | 6.532000 | 0.596000 | 1.596000 |
+GRB230815A   |   1185505    | 713789427.776000 | -14.776000 | 9.224000 | -11.776000 | 5.224000 | -0.776000 | 1.224000 | 0.224000 | 1.224000 |
+GRB230805B   |   1183217    | 712927462.848000 | -17.656000 | 109.308000 | -11.752000 | 102.768000 | 6.740000 | 45.276000 | -17.664000 | -16.664000 |
+GRB230802A   |   1182085    | 712651857.920000 | -6.040000 | 230.344000 | -1.780000 | 173.372000 | 4.976000 | 104.436000 | 0.604000 | 1.604000 |
+GRB230728A   |   1181187    | 712205434.944000 | -5.620000 | 10.740000 | -4.852000 | 9.908000 | 0.364000 | 5.912000 | 0.408000 | 1.408000 |
+GRB230723B   |   1180410    | 711805384.576000 | -0.216000 | 7.936000 | 0.116000 | 5.872000 | 1.240000 | 3.912000 | 1.636000 | 2.636000 |
+GRB230628E   |   1177054    | 709684002.240000 | -0.504000 | 4.808000 | -0.328000 | 4.392000 | 0.632000 | 2.664000 | 0.332000 | 1.332000 |
+GRB230622A   |   1176135    | 709121092.800000 | -0.032000 | 24.160000 | 0.160000 | 23.712000 | 0.544000 | 13.728000 | 0.076000 | 1.076000 |
+GRB230618A   |   1175438    | 708753104.000000 | -27.784000 | 107.004000 | -18.248000 | 85.804000 | -8.660000 | 34.140000 | 9.040000 | 10.040000 |
+GRB230606A   |   1173731    | 707703308.736000 | -0.264000 | 16.968000 | 0.088000 | 16.344000 | 1.496000 | 14.504000 | 0.924000 | 1.924000 |
+GRB230510A   |   1167973    | 705413219.840000 | -145.632000 | 56.096000 | -138.976000 | 39.392000 | -1.376000 | 21.344000 | 15.884000 | 16.884000 |
[…]
