+++ pyproject.toml
@@ -41,17 +41,17 @@ classifiers = [
     "Operating System :: POSIX :: Linux"
 ]
 dependencies = [
-    "fabada==0.1",
-    "matplotlib==3.7.1",
-    "moviepy==1.0.3",
-    "numpy==1.24.2",
-    "openTSNE==0.7.1",
-    "pandas==2.0.0",
-    "requests==2.31.0",
-    "scipy==1.10.1",
-    "tqdm==4.65.0",
-    "tables==3.8.0",
-    "scikit-learn==1.2.2"
+    "fabada>=0.1",
+    "matplotlib>=3.7.1",
+    "moviepy>=1.0.3",
+    "numpy>=1.24.2",
+    "openTSNE>=0.7.1",
+    "pandas>=2.0.0",
+    "requests>=2.31.0",
+    "scipy>=1.10.1",
+    "tqdm>=4.65.0",
+    "tables>=3.8.0",
+    "scikit-learn>=1.2.2"
 ]
 
 [project.urls]
