### 2023-12-05 #5923 @KenethGarcia
@dfm  Thank you for reaching out! We appreciate your proactive approach to the ongoing review. Actually, we are addressing the issues you've opened, as well as carefully considering @wkerzendorf  comments. We expect to answer the issues in the next days for further comments. Actually, there are minor updates in the GitHub repository, and we are implementing a conda installation with an updated catalogue of GRBs (up to 2023).
