### 2023-12-05 KenethGarcia/ClassiPyGRB#2 @KenethGarcia
Thank you for your comment. The execution time of the light curve limiter function depends on the query time of the data. The 9 seconds mentioned in the tutorials reflect a scenario where data was previously downloaded locally. However, if **ClassiPyGRB** retrieves data directly from the [Swift/BAT website](https://swift.gsfc.nasa.gov/results/batgrbcat/) to limit light curves, the duration can significantly vary based on internet connection speed and server availability.

We have now included an explanation of this in the tutorials, along with a strong recommendation to download local data if you aim to limit the dataset to a few seconds.
