+++ README.md
@@ -2,8 +2,15 @@
 
 **ClassiPyGRB** is a Python 3 package to download, process, visualize and classify Gamma-Ray-Bursts (GRBs) from the Swift/BAT Telescope (https://swift.gsfc.nasa.gov/about_swift/bat_desc.html) database. It is distributed over the GNU General Public License Version 2 (1991). Please, read the complete description of the method and its application to GRBs in this [publication](JOSS_Docs/paper.md).
 
+# Statement of need
+The Swift Burst Alert Telescope (BAT) is a wide-field, hard Gamma-ray detector that, since its launch in 2004, has played an important role in the high-energy astrophysical field. A preliminary query on the Astrophysics Data System (ADS) with the keyword "Swift/BAT" indicates that over 7000 research works have been uploaded to its platform (up to December 2023). Furthermore, recently has been increased the number of studies per year, evidencing the relevance and impact of the use of the Swift/BAT database.
+
+Although the Swift/BAT database is [publicly available](https://swift.gsfc.nasa.gov/results/batgrbcat/), for first-time users it might be a challenge to download and process the data. The data is stored in multiple directories, depending on the GRB. Moreover, the data is not pre-processed, and the user must perform the data manipulation and interpolation by themselves. These issues make the data processing a time-consuming task. **ClassiPyGRB** is a Python 3 package that aims to solve these problems by providing a simple and intuitive interface to download, process, visualize, and classify the photometric data of GRBs from the Swift/BAT database.
+
+**ClassiPyGRB** allows researchers to query light curves for any GRB in the Swift/BAT database simply and intuitively. The package also provides a set of tools to preprocess the data, including noise/duration reduction and interpolation. Moreover, The package also provides a set of facilities and tutorials to classify GRBs based on their light curves, following [Jespersen et al.(2020)](https://doi.org/10.3847/2041-8213/ab964d) and [Garcia-Cifuentes et al.(2023)](https://doi.org/10.3847/1538-4357/acd176). This method is based on a dimensionality reduction of the data using t-Distributed Stochastic Neighbour Embedding (TSNE), where the user can visualize the results using a Graphical User Interface (GUI). The user can also plot and animate the results of the TSNE analysis, allowing to perform a deeper hyperparameter grid search. The package is distributed over the GNU General Public Licence Version 2 (1991).
+
 # Attribution
-If you use this code in a publication, please refer to the package by its name and cite [Garcia-Cifuentes et al.(2023)](https://ui.adsabs.harvard.edu/abs/2023arXiv230408666G/abstract) -> [arXiv link](https://arxiv.org/abs/2304.08666). Any question, please email [Keneth Garcia-Cifuentes](mailto:kenet.garcia@correo.nucleares.unam.mx).
+If you use this code in a publication, please refer to the package by its name and cite [Garcia-Cifuentes et al.(2023)](https://doi.org/10.3847/1538-4357/acd176) -> [Astrophysical Journal Vol. 951 No. 1](https://doi.org/10.3847/1538-4357/acd176). Any question, please email [Keneth Garcia-Cifuentes](mailto:kenet.garcia@correo.nucleares.unam.mx).
 
 ## Dependencies
 This repository requires Python 3.8 or high, and a list of packages downloaded automatically ([numpy](https://github.com/numpy/numpy), [scikit-learn](https://scikit-learn.org/stable/index.html), etc). In addition, it is required to install all the dependencies related to Tkinter, Pillow, and ImageTK. In Debian-based distros you can install these packages by running the following commands:
