### 2023-12-12 KenethGarcia/ClassiPyGRB#3 @KenethGarcia
Thank you for the comment. We already added a section in the README called "Statement of need" where it is specified:

- An introduction to the problem/challenge
- What are the current limitations or gaps
- Benefits of the package
- Relevance to the Community and target audience
- Potential impact
- User scenarios or use cases
