### 2023-12-20 aebc51f
Added conda/mamba installation instructions and some comments about PyTables #1
