### 2023-12-20 KenethGarcia/ClassiPyGRB#1 @KenethGarcia
Thank you for your comment! We want to highlight the following points for this issue:

1. We bounded all the requirements for **ClassiPyGRB**. As you mentioned, there is no reason to pin any package. We ran all the tests and tutorials using the latest versions (up to Dec. 20th), and there are no issues to fix.
2. We updated the installation tutorial by adding conda/mamba installation instructions. We want to stress that we conducted several hours of testing and environment validation to ensure a seamless and reliable installation process. We remark also that in previously installed tkinter/pytables OS, the requirements creating the conda environment are unnecessary.
3. As the [PyTables documentation states](https://www.pytables.org/usersguide/installation.html), the need for the `tables` requirement for the package was already established as a dependency for installations with pip (see the [TOML file](https://github.com/KenethGarcia/ClassiPyGRB/blob/master/pyproject.toml)). However, we mention to install PyTables in conda/mamba environments on the README.
