### 2023-12-22 97f0ca3
Added a more readable installation instructions #1
