+++ .github/workflows/tests.yml
@@ -13,7 +13,6 @@ on:
     paths:
       - src/**
       - .github/workflows/tests.yml
-    workflow_dispatch:
 
 jobs:
   test:
