+++ .github/workflows/tests.yml
@@ -4,15 +4,9 @@ on:
   push:
     branches:
       - main
-    paths:
-      - src/**
-      - .github/workflows/tests.yml
   pull_request:
     branches:
       - main
-    paths:
-      - src/**
-      - .github/workflows/tests.yml
 
 jobs:
   test:
