+++ .github/workflows/tests.yml
@@ -3,10 +3,16 @@ name: Run Tests
 on:
   push:
     branches:
-      - main
+      - master
+    paths:
+      - src/**
+      - .github/workflows/tests.yml
   pull_request:
     branches:
-      - main
+      - master
+    paths:
+      - src/**
+      - .github/workflows/tests.yml
 
 jobs:
   test:
