+++ .github/workflows/tests.yml
@@ -40,7 +40,6 @@ jobs:
       - name: Install Dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
 
       - name: Install package
         shell: bash -l {0}
