### 2023-12-22 eaca363
Removed use of TkAgg
