+++ .github/workflows/tests.yml
@@ -20,7 +20,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10"]
+        python-version: "3.10"
 
     steps:
       - name: Checkout Repository
