### 2024-01-10 KenethGarcia/ClassiPyGRB#5 @KenethGarcia
Thank you for bringing up the concern regarding the tests in the repository. I'm pleased to inform you that we have implemented the following solutions:

1. **Documentation for Test Running Procedure:**
   We have added a documentation section in the repository's README, explaining how to run the tests.

2. **GitHub Actions Workflow for Test Execution:**
   A GitHub Actions workflow has been created to automate the testing process. This workflow is triggered on each push to the repository, and it ensures that the tests are executed in a controlled and consistent environment. You can find the workflow configuration file in the `.github/workflows` directory.

Thank you for your valuable input and collaboration!
