### 2024-01-24 #5923 @KenethGarcia
Dear @dfm and @wkerzendorf,

We are writing to leave you with some updates on changes that have been made to **ClassiPyGRB** based on your comments and feedback:

### Regarding the comment about “Installation and cross-platform support”:
We updated the installation tutorial by adding conda/mamba installation instructions. We conducted several hours of testing and environment validation to ensure a seamless and reliable installation process. Furthermore, we also remark that in previously installed tkinter/pytables OS, the requirements of creating the conda environment are unnecessary.

### About testing and CI workflows:
We have implemented the following documentation for Test Running Procedure:

1. We have added a documentation section in the repository's README, explaining how to run the tests.
2. GitHub Actions Workflow for Test Execution:
3. A GitHub Actions workflow has been created to automate the testing process. This workflow is triggered on each push to the repository, and it ensures that the tests are executed in a controlled and consistent environment. You can find the workflow configuration file in the .github/workflows directory.

### About the Statement of Need:
We added a new section where we specify:

- An introduction to the problem/challenge
- What are the current limitations or gaps
- The benefits of the package
- The relevance to the Community and target audience
- The potential impact
- User scenarios or use cases

Thank you very much for your comments, we will be looking forward to any additional suggestions to this library.

Best Regards
