### 2024-02-11 KenethGarcia/ClassiPyGRB#1 @dfm
Sorry for the delay - the updated installation instructions look great!
