### 2024-02-11 KenethGarcia/ClassiPyGRB#2 @dfm
This makes sense - thanks for clearing it up in the tutorial!
