### 2024-02-11 KenethGarcia/ClassiPyGRB#3 @dfm
This is excellent!
