### 2024-02-11 KenethGarcia/ClassiPyGRB#5 @dfm
This is great! I can confirm that I can now easily run all the tests locally and they all pass.
