### 2024-02-11 KenethGarcia/ClassiPyGRB#6 @dfm
Hardcoded absolute paths in tutorials
In several tutorials (including the TSNE intro), there are hard coded paths of the form `/home/keneth/Documents/ClassiPyGRB/docs/...` used. While it is noted that these should be changed, it seems like it would also be fine to use relative paths (e.g. `animation1.gif`), and I think that would be more generally useful! Not a major blocker, but I always recommend that full paths like these be removed from any code/tutorial that's meant for public consumption.

ref: https://github.com/openjournals/joss-reviews/issues/5923
