### 2024-02-11 KenethGarcia/ClassiPyGRB#4 @dfm
I'm going to check off this entry as complete in the JOSS review since there are good docstrings, but I think it would be great to expose this more clearly in the docs!
