### 2024-02-11 90d787d
Small edits to JOSS paper
