### 2024-02-18 61595be
Removed hardcoded paths creating animations #6
