### 2024-02-18 KenethGarcia/ClassiPyGRB#6 @KenethGarcia
Thank you for the comment!

Of course, we have not noticed the hardcoded paths in the Tutorial of TSNE Overview. We have changed these paths for simplicity. Moreover, I want to stress that in **ClassiPyGRB**, the saved filename on the *tsne_animation* instance can be any particular or relative path.
