### 2024-02-22 #5923 @wkerzendorf
Can you generate new proofs?
