### 2024-03-07 #5923 @wkerzendorf
The `State of the Field` in the paper is missing. Also there are references to a `requeriments.txt` in the paper doesn't seem to exist in the code.
