### 2024-03-15 7761531
Corrected requirements mention on JOSS Paper
