+++ JOSS_Docs/paper.bib
@@ -101,3 +101,59 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220105145S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Angulo-Valdez:2024,
+       author = {{Angulo-Valdez}, Camila and {Becerra}, Rosa L. and {Pereyra}, Margarita and {Garcia-Cifuentes}, Keneth and {Vargas}, Felipe and {Watson}, Alan M. and {De Colle}, Fabio and {Fraija}, Nissim and {Butler}, Nathaniel R. and {Dainotti}, Maria G. and {Dichiara}, Simone and {Lee}, William H. and {Troja}, Eleonora and {Bloom}, Joshua S. and {Gonz{\'a}lez}, J. Jes{\'u}s and {Kutyrev}, Alexander S. and {Prochaska}, J. Xavier and {Ramirez-Ruiz}, Enrico and {Richer}, Michael G.},
+        title = "{Machine-learning enhanced photometric analysis of the extremely bright GRB 210822A}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {gamma-ray burst: individual: GRB 210822A, gamma-ray bursts: general, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2024,
+        month = jan,
+       volume = {527},
+       number = {3},
+        pages = {8140-8150},
+          doi = {10.1093/mnras/stad3624},
+archivePrefix = {arXiv},
+       eprint = {2309.10106},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024MNRAS.527.8140A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Parsotan:2023,
+       author = {{Parsotan}, Tyler and {Laha}, Sibasish and {Palmer}, David M. and {Lien}, Amy and {Cenko}, S. Bradley and {Krimm}, Hans and {Markwardt}, Craig},
+        title = "{BatAnalysis: A Comprehensive Python Pipeline for Swift BAT Survey Analysis}",
+      journal = {The Astrophysical Journal},
+     keywords = {Astronomy data analysis, Astronomy data acquisition, Open source software, 1858, 1860, 1866, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2023,
+        month = aug,
+       volume = {953},
+       number = {2},
+          eid = {155},
+        pages = {155},
+          doi = {10.3847/1538-4357/ace325},
+archivePrefix = {arXiv},
+       eprint = {2303.06255},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...953..155P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{DeLaunay:2022,
+       author = {{DeLaunay}, James and {Tohuvavohu}, Aaron},
+        title = "{Harvesting BAT-GUANO with NITRATES (Non-Imaging Transient Reconstruction and Temporal Search): Detecting and Localizing the Faintest Gamma-Ray Bursts with a Likelihood Framework}",
+      journal = {The Astrophysical Journal},
+     keywords = {Gamma-ray bursts, Space telescopes, Gravitational wave sources, 629, 1547, 677, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
[…]
