### 2024-03-15 #5923 @KenethGarcia
Dear @wkerzendorf,

We have corrected the references to a `requirements.txt` of ClassiPyGRB. In the new version, we specify that all the necessary packages will be automatically handled during installation.

On the other hand, we have updated the Statement of Need section, adding a more accurate description of what problems the software is designed to solve, who the target audience is, and the differences with other similar packages.

Thank you very much for your comments again, we will be looking forward to any additional suggestions to this package.

Best regards
