### 2024-03-21 #5923 @KenethGarcia
Hi @wkerzendorf,

Just a quick follow-up regarding the previous feedback:

Was the addition of the State of the Field section and the adjustments to the references satisfactory?

Thank you!

Best regards,

Keneth
