### 2024-03-26 #5923 @KenethGarcia
Dear @dfm ,

I hope this message finds you well. I am writing to seek your guidance and assistance regarding the recent feedback we received on ClassiPyGRB.

As per the suggestions provided by @elotroreferee, we have made the corrections required, including addressing the missing State of the Field section and rectifying the references to a requirements.txt file. However, there has been no confirmation on whether these adjustments meet the JOSS requirements. Can u help us?

Thank you for your attention to this matter, and I look forward to your prompt response.

Best regards,

Keneth
On behalf of the authors
