### 2024-04-02 #5923 @KenethGarcia
Dear dfm,

As you request, we have added:

1. A new version in the GitHub repository of ClassiPyGRB, with the label `v1.1.0`
2. We have reviewed the paper again looking for updates, and we have made minor changes in the affiliations and acknowledgments sections.
3. We have updated an archived version of the software in Zenodo, and it has assigned the DOI [10.5281/zenodo.10909942](https://zenodo.org/doi/10.5281/zenodo.10909942). Moreover, we also have carefully checked the title and authors.

Thank you for the indications and we will stay in contact,

Best regards,

Keneth
On behalf of the authors
